+++ tests/Functional/Command/Ingester/IngesterCommandTest.php
@@ -107,10 +107,11 @@ class IngesterCommandTest extends KernelTestCase
             $this->normalizeDisplay($this->commandTester->getDisplay())
         );
 
+        $messageError = 'Argument 1 passed to OAT\SimpleRoster\Entity\Infrastructure::setLtiSecret() '
+            . 'must be of the type string, null given';
+
         self::assertStringContainsString(
-            <<<MESSAGE_ERROR
-Argument 1 passed to OAT\SimpleRoster\Entity\Infrastructure::setLtiSecret() must be of the type string, null given
-MESSAGE_ERROR,
+            $messageError,
             $this->commandTester->getDisplay()
         );